requirements.txt
praw>=7.7.1,<8
supabase>=2.0.0,<3
python-dotenv>=1.0.0,<2
urllib3<2


